+++ src/interfaces/rtc_peer_connection/create_session_description_observer.cc
@@ -15,7 +15,7 @@
 #include "src/dictionaries/node_webrtc/some_error.h"
 
 void node_webrtc::CreateSessionDescriptionObserver::OnSuccess(webrtc::SessionDescriptionInterface* description) {
-  auto maybeDescription = node_webrtc::From<RTCSessionDescriptionInit>(const_cast<const webrtc::SessionDescriptionInterface*>(description));
+  auto maybeDescription = node_webrtc::From<RTCSessionDescriptionInit>(description);
   delete description;
   if (maybeDescription.IsInvalid()) {
     Reject(node_webrtc::SomeError(maybeDescription.ToErrors()[0]));