+++ src/node/error_factory.h
@@ -31,19 +31,19 @@ class ErrorFactory {
 
   static void Init(Napi::Env, Napi::Object);
 
-  static Napi::Value CreateError(Napi::Env, std::string);
-  static Napi::Value CreateInvalidAccessError(Napi::Env, std::string);
-  static Napi::Value CreateInvalidModificationError(Napi::Env, std::string);
-  static Napi::Value CreateInvalidStateError(Napi::Env, std::string);
-  static Napi::Value CreateNetworkError(Napi::Env, std::string);
-  static Napi::Value CreateOperationError(Napi::Env, std::string);
-  static Napi::Value CreateRangeError(Napi::Env, std::string);
-  static Napi::Value CreateSyntaxError(Napi::Env, std::string);
+  static Napi::Value CreateError(Napi::Env, std::string const&);
+  static Napi::Value CreateInvalidAccessError(Napi::Env, std::string const&);
+  static Napi::Value CreateInvalidModificationError(Napi::Env, std::string const&);
+  static Napi::Value CreateInvalidStateError(Napi::Env, std::string const&);
+  static Napi::Value CreateNetworkError(Napi::Env, std::string const&);
+  static Napi::Value CreateOperationError(Napi::Env, std::string const&);
+  static Napi::Value CreateRangeError(Napi::Env, std::string const&);
+  static Napi::Value CreateSyntaxError(Napi::Env, std::string const&);
 
  private:
   static Napi::Value SetDOMException(const Napi::CallbackInfo&);
   static const char* DOMExceptionNameToString(DOMExceptionName);
-  static Napi::Value CreateDOMException(Napi::Env, std::string, DOMExceptionName);
+  static Napi::Value CreateDOMException(Napi::Env, std::string const&, DOMExceptionName);
   static Napi::FunctionReference& _DOMException();
 };